+++ packages/rocket_client/CHANGELOG.md
@@ -3,6 +3,6 @@
 [CHANGELOG](https://github.com/JahezAcademy/flutter_rocket/blob/dev/packages/flutter_rocket/CHANGELOG.md)
 
 ## 0.0.2
-    - Added `onResponse` parameter to `RocketRequest` for handle common cases
 
-
+* Added `onResponse` parameter to `RocketRequest` for handle common cases
+* Added onError method parameter to request method