+++ packages/rocket_client/CHANGELOG.md
@@ -3,7 +3,9 @@
 * [Old changelogs](https://pub.dev/packages/flutter_rocket/changelog)
 
 # 0.0.2
+
 * Renamed `targetData` param to `target` on request `method`
 * Fixed issues on `inspect` & `targetData` param of `request`
-* Return `RocketModel` from `request` method instead of dynamic
+* Return `RocketModel` from `request` method instead of `dynamic`
 * Added `onResponse` parameter to `RocketRequest` for handle common cases
+* Added `onError` method parameter to `request` method